+++ ikarus/assembler/assemblermanipulatorfuser.hh
@@ -82,6 +82,7 @@ public:
   using WrappedAssembler::boundToDBCOption;
   using WrappedAssembler::boundToRequirement;
   using WrappedAssembler::createFullVector;
+  using WrappedAssembler::finiteElements;
 
   using WrappedAssembler::affordanceCollection;
   using WrappedAssembler::constraintsBelow;
@@ -173,6 +174,7 @@ public:
   using WrappedAssembler::boundToDBCOption;
   using WrappedAssembler::boundToRequirement;
   using WrappedAssembler::createFullVector;
+  using WrappedAssembler::finiteElements;
 
   using WrappedAssembler::affordanceCollection;
   using WrappedAssembler::constraintsBelow;
@@ -283,6 +285,7 @@ public:
   using WrappedAssembler::boundToDBCOption;
   using WrappedAssembler::boundToRequirement;
   using WrappedAssembler::createFullVector;
+  using WrappedAssembler::finiteElements;
 
   using WrappedAssembler::affordanceCollection;
   using WrappedAssembler::constraintsBelow;