+++ ikarus/solver/nonlinearsolver/solverinfos.hh
@@ -3,10 +3,12 @@
 
 /**
  * \file solverinfos.hh
- * \brief Implementation of the Newton-Raphson method for solving nonlinear equations.
+ * \brief Implementation of the solver information returned by the nonlinear solvers.
  */
 
 #pragma once
+
+#include <limits>
 
 namespace Ikarus {
 /**